requirements.txt
requests>=2.4.0
websocket-client>=0.22.0,<=0.44.0
slacker>=0.9.50
six>=1.10.0
numpy
pandas
piazza-api
pytz
slackclient